+++ webpack.production.config.js
@@ -29,13 +29,13 @@ module.exports = {
       },
       {
           test: /\.css$/,
-          include: /node_modules/,
+          include:  [/node_modules/, /styles_custom/],
           loader: 'style-loader!css-loader'
       },
       {
         test: /\.css$/,
-        exclude: /node_modules/,
-        loader: 'style!css?modules!postcss'
+        exclude: [/node_modules/, /styles_custom/],
+        loader: 'style-loader!css-loader?modules'
       },
       { test: /\.(png|jpg)$/,
         loader: 'url-loader?limit=8192'